Move Output module inside CLI
