--- lib/cobra_commander/output.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-require_relative "output/flat_list"
-require_relative "output/ascii_tree"
-require_relative "output/dot_graph"
-require_relative "output/interactive_printer"
-require_relative "output/markdown_printer"
