+++ spec/cobra_commander/cli/output/change_spec.rb
@@ -1,20 +1,20 @@
 # frozen_string_literal: true
 
 require "spec_helper"
-require "cobra_commander/change"
+require "cobra_commander/cli/output/change"
 
-RSpec.describe CobraCommander::Change do
+RSpec.describe CobraCommander::CLI::Output::Change do
   let(:umbrella) { fixture_umbrella("app") }
 
   it "successfully instantiates" do
-    expect(CobraCommander::Change.new(umbrella, "full", "master")).to be_truthy
+    expect(CobraCommander::CLI::Output::Change.new(umbrella, "full", "master")).to be_truthy
   end
 
   describe ".run!" do
     context "with json results" do
       context "with no changes" do
         it "lists no files" do
-          changes = CobraCommander::Change.new(umbrella, "json", "master", changes: [])
+          changes = CobraCommander::CLI::Output::Change.new(umbrella, "json", "master", changes: [])
 
           expect { changes.run! }.to output(<<~OUTPUT
             {"changed_files":[],"directly_affected_components":[],"transitively_affected_components":[],"test_scripts":[],"component_names":[],"languages":{"ruby":false,"javascript":false}}
@@ -25,7 +25,8 @@ RSpec.describe CobraCommander::Change do
 
       context "with a change inside a component" do
         it "lists change, affected component, and test" do
-          change = CobraCommander::Change.new(umbrella, "json", "master", changes: ["#{umbrella.path}/components/a"])
+          change = CobraCommander::CLI::Output::Change.new(umbrella, "json", "master",
+                                                           changes: ["#{umbrella.path}/components/a"])
 
           expect { change.run! }.to output(<<~OUTPUT
             {"changed_files":["#{umbrella.path}/components/a"],"directly_affected_components":[{"name":"a","path":["#{umbrella.path}/components/a"],"type":["bundler"]}],"transitively_affected_components":[],"test_scripts":["#{umbrella.path}/components/a/test.sh"],"component_names":["a"],"languages":{"ruby":true,"javascript":false}}
@@ -36,7 +37,8 @@ RSpec.describe CobraCommander::Change do
 
       context "with change inside a very utilized component" do
         it "lists changes, affected components, and tests" do
-          change = CobraCommander::Change.new(umbrella, "json", "master", changes: ["#{umbrella.path}/components/e"])
+          change = CobraCommander::CLI::Output::Change.new(umbrella, "json", "master",
+                                                           changes: ["#{umbrella.path}/components/e"])
 
           expect { change.run! }.to output(<<~OUTPUT
             {"changed_files":["#{umbrella.path}/components/e"],"directly_affected_components":[{"name":"e","path":["#{umbrella.path}/components/e"],"type":["yarn"]}],"transitively_affected_components":[{"name":"a","path":["#{umbrella.path}/components/a"],"type":["bundler"]},{"name":"b","path":["#{umbrella.path}/components/b"],"type":["yarn","bundler"]},{"name":"c","path":["#{umbrella.path}/components/c"],"type":["bundler"]},{"name":"d","path":["#{umbrella.path}/components/d"],"type":["bundler"]},{"name":"f","path":["#{umbrella.path}/components/f"],"type":["yarn"]},{"name":"g","path":["#{umbrella.path}/components/g"],"type":["yarn"]},{"name":"h","path":["#{umbrella.path}/components/h"],"type":["yarn","bundler"]},{"name":"node_manifest","path":["#{umbrella.path}/node_manifest"],"type":["yarn"]}],"test_scripts":["#{umbrella.path}/components/a/test.sh","#{umbrella.path}/components/b/test.sh","#{umbrella.path}/components/c/test.sh","#{umbrella.path}/components/d/test.sh","#{umbrella.path}/components/e/test.sh","#{umbrella.path}/components/f/test.sh","#{umbrella.path}/components/g/test.sh","#{umbrella.path}/components/h/test.sh","#{umbrella.path}/node_manifest/test.sh"],"component_names":["a","b","c","d","e","f","g","h","node_manifest"],"languages":{"ruby":true,"javascript":true}}
@@ -49,7 +51,7 @@ RSpec.describe CobraCommander::Change do
     context "with full results" do
       context "with no changes" do
         it "lists no files" do
-          changes = CobraCommander::Change.new(umbrella, "full", "master", changes: [])
+          changes = CobraCommander::CLI::Output::Change.new(umbrella, "full", "master", changes: [])
 
           expect { changes.run! }.to output(<<~OUTPUT
             <<< Changes since last commit on master >>>
@@ -66,7 +68,7 @@ RSpec.describe CobraCommander::Change do
 
       context "with a change outside a component" do
         it "just lists single change" do
-          change = CobraCommander::Change.new(umbrella, "full", "master", changes: ["/change"])
+          change = CobraCommander::CLI::Output::Change.new(umbrella, "full", "master", changes: ["/change"])
 
           expect { change.run! }.to output(<<~OUTPUT
             <<< Changes since last commit on master >>>
@@ -84,7 +86,8 @@ RSpec.describe CobraCommander::Change do
 
       context "with a change inside a component" do
         it "lists change, affected component, and test" do
-          change = CobraCommander::Change.new(umbrella, "full", "master", changes: ["#{umbrella.path}/components/a"])
+          change = CobraCommander::CLI::Output::Change.new(umbrella, "full", "master",
+                                                           changes: ["#{umbrella.path}/components/a"])
 
           expect { change.run! }.to output(<<~OUTPUT
             <<< Changes since last commit on master >>>
@@ -104,7 +107,7 @@ RSpec.describe CobraCommander::Change do
 
       context "with changes inside multiple components" do
         it "lists changes, affected components, and tests" do
-          change = CobraCommander::Change.new(
+          change = CobraCommander::CLI::Output::Change.new(
             umbrella, "full", "master",
             changes: ["#{umbrella.path}/components/a", "#{umbrella.path}/components/b"]
           )
@@ -143,7 +146,8 @@ RSpec.describe CobraCommander::Change do
 
       context "with change inside a very utilized component" do
         it "lists changes, affected components, and tests" do
-          change = CobraCommander::Change.new(umbrella, "full", "master", changes: ["#{umbrella.path}/components/e"])
+          change = CobraCommander::CLI::Output::Change.new(umbrella, "full", "master",
+                                                           changes: ["#{umbrella.path}/components/e"])
 
           expect { change.run! }.to output(<<~OUTPUT
             <<< Changes since last commit on master >>>
