+++ spec/cobra_commander/cli/output/dot_graph_spec.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
 require "spec_helper"
-require "cobra_commander/output/dot_graph"
+require "cobra_commander/cli/output/dot_graph"
 
-RSpec.describe CobraCommander::Output::DotGraph do
+RSpec.describe CobraCommander::CLI::Output::DotGraph do
   describe "#generate!" do
     let(:expected_dot) { fixture_file("expected.dot") }
     let(:generated_dot) { StringIO.new }
@@ -22,7 +22,7 @@ RSpec.describe CobraCommander::Output::DotGraph do
     end
 
     it "correctly generates graph.dot" do
-      CobraCommander::Output::DotGraph.generate(umbrella.root, generated_dot)
+      CobraCommander::CLI::Output::DotGraph.generate(umbrella.root, generated_dot)
 
       expect(generated_dot.string).to eql expected_dot.read
     end
